Refactor: Moves calculateFieldStats, LogsParsers and getParser to grafana/data
